[VB-22] refactoring
